+++ packages/gamut-labs/src/AppHeaderMobile/AppHeaderSubMenuTarget/index.tsx
@@ -2,7 +2,7 @@ import { Anchor, Text } from '@codecademy/gamut';
 import { ArrowChevronRightIcon } from '@codecademy/gamut-icons';
 import { css } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
-import React from 'react';
+import * as React from 'react';
 
 import {
   AppHeaderCatalogDropdownItem,