+++ packages/gamut/src/Form/FormError.tsx
@@ -1,7 +1,8 @@
 import { theme, variant } from '@codecademy/gamut-styles';
 import { StyleProps } from '@codecademy/variance';
 import styled from '@emotion/styled';
-import React, { HTMLAttributes } from 'react';
+import { HTMLAttributes } from 'react';
+import * as React from 'react';
 
 const errorSpanVariants = variant({
   base: {