+++ packages/gamut/src/GridForm/GridFormButtons/index.tsx
@@ -1,5 +1,6 @@
 import { ColorModes } from '@codecademy/gamut-styles';
-import React, { ComponentProps } from 'react';
+import { ComponentProps } from 'react';
+import * as React from 'react';
 
 import { Box } from '../../Box';
 import { CTAButton, FillButton, TextButton } from '../../Button';