+++ packages/gamut/src/Pagination/utils.tsx
@@ -1,5 +1,6 @@
 import { AnimatePresence, motion } from 'framer-motion';
-import React, { useEffect, useRef } from 'react';
+import { useEffect, useRef } from 'react';
+import * as React from 'react';
 
 import { BaseEllipsisButton } from './EllipsisButton';
 import { PaginationButton } from './PaginationButton';