+++ packages/gamut/src/ProgressBar/index.tsx
@@ -1,7 +1,7 @@
 import { PatternProps } from '@codecademy/gamut-patterns';
 import { variant } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
-import React from 'react';
+import * as React from 'react';
 
 import { Text } from '../Typography';