+++ packages/gamut/src/typings/react.d.ts
@@ -1,4 +1,5 @@
-import React, { Attributes, ReactHTML } from 'react';
+import { Attributes, ReactHTML } from 'react';
+import * as React from 'react';
 
 /**
  * Simple tag string, component class, or functional component for a render prop.